requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-JWT-Extended==3.18.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==1.4.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.21
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
razorpay